+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 
 *   Support invoking a constant macro as a function name.
 *   If a function's argument pattern looks like a keyword block, it tried to codegen as such. Fixed.
+*   If an Erlang variable name was a capitalized version of an Elixir keyword (such as "End"), it would generate uncompilable Elixir code. Fixed.
+*   Start some optional integration tests that convert and run tests against common Erlang libraries.
 
 ## v0.0.6 (2016-01-19)